Mon, Jul 17, 2023, 4:01 AM +08:00

+++ finance/app.py
@@ -45,13 +45,13 @@ def buy():
     """Buy shares of stock"""
     if request.method == "POST":
         symbol = request.form.get("symbol")
-        shares = request.form.get("shares")
+        shares = int(request.form.get("shares"))
         symbol_dict = lookup(symbol)
         user_id = session["user_id"]
         username = db.execute("SELECT username FROM users WHERE id = ?", user_id)
-        cash = db.execute("SELECT cash FROM users WHERE id = ?", user_id)
-        price = symbol_dict["price"]
-        total_price = price * shares
+        cash = int(db.execute("SELECT cash FROM users WHERE id = ?", user_id))
+        price = int(symbol_dict["price"])
+        total_price = int(price * shares)
 
         if symbol and shares:
             if symbol_dict:
